Base: [FEAT] implemented run_one
run_one executes a command in a subprocess. It makes sure only one such
command is executed at a time. This is achieved by checking the provided
PID file and killing the process with id found in this file. The PID of
the new process is written to the PID file.