[FEAT] resume now respects "after"
You can now request, that the resume command is executed after a
delay. This delay can be given in the message as hash key
"after". Furthermore, a default value can be set in the
config. Tapper::Action forks a child process that sleeps the "after"
time so normal operation is not stalled while waiting.